First set of large changes for VIF. Not completely stable on edge cases.
VIF estimation of r.squared included. Several other updates for edge cases. Not perfectly stable. On the read estate data it can still happen that all experts are finished and haven't started new passes. Updating to a different algorithm, where store results of previous tests instead of just where to move to. The idea will be to cycle quickly through skip-able tests, not to literally skip to the next test which rejects.